extract setupMultistoreRoutes tests
